refactor: bump up fixtures to global conftest
